+++ src/util/mod.rs
@@ -3,3 +3,4 @@ pub mod non_macro;
 pub mod useful;
 pub mod api;
 pub mod pickle;
+pub mod writer;